Improved altitude control during F_LAND option. Eliminated the PITCH parameter during landing, found a much better way to do it. Tests show that it is really slick:
With the throttle off, the altitude control follows the line between waypoints, with the throttle off. To keep the plane from stalling with the throttle off, there is one important change from how altitude pitch control works in normal flight: positive pitch requests are ignored, only negative ones are. This places a ceiling on the plane, and it is steadily drawn down to the runway, landing pretty close to where you want it. Glide slope and landing point are defined by the waypoints. The landing target waypoint should be unreachable, and below ground.